requirements.txt
Django==1.11.4
django-rest-framework==0.1.0
djangorestframework==3.6.3
gunicorn==19.7.1
PyJWT==1.5.2
python-decouple==3.0
pytz==2017.2
